Move password hashing and verification to security file
