+++ api/security.py
@@ -2,6 +2,18 @@ import os
 from datetime import datetime, timedelta
 
 import jwt
+from passlib.context import CryptContext
+
+pw_ctx = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+
+def hash_password(plain_password: str) -> str:
+    return pw_ctx.hash(plain_password)
+
+
+def verify_password_hash(plain_password: str, hashed_password: str):
+    return pw_ctx.verify(plain_password, hashed_password)
+
 
 SECRET_KEY = str(os.getenv("JWT_SECRET_KEY"))
 ALGORITH = str(os.getenv("JWT_ALGORITHM"))
